Refactor application key mapping and application management endpoints
