+++ portals/developer-portal/src/scripts/oauth2-key-generation.js
@@ -77,9 +77,6 @@ async function generateApplicationKey(formId, appId, keyType, keyManager, client
         grantTypes = jsonObject.grantTypes;
     }
     const payload = JSON.stringify({
-        "applicationName": clientName,
-        "apis": apiList,
-        "tokenType": "OAUTH",
         "tokenDetails": {
             "grantTypesToBeSupported": grantTypes,
             "keyType": keyType,
@@ -94,7 +91,7 @@ async function generateApplicationKey(formId, appId, keyType, keyManager, client
         "clientID": document.getElementById("clientIDInput-" + keyType)?.value?.trim(),
     })
     try {
-        const response = await fetch(`/devportal/organizations/${orgID}/app-key-mapping`, {
+        const response = await fetch(`/devportal/applications/${appId}/generate-keys`, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
